+++ packages/astro-integration/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sylphx/silk-astro
 
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @sylphx/silk@1.2.0
+  - @sylphx/silk-react@1.0.3
+
 ## 2.0.1
 
 ### Patch Changes