+++ packages/preset-material/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sylphx/silk-preset-material
 
+## 3.0.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @sylphx/silk@1.2.0
+
 ## 2.0.0
 
 ### Major Changes