+++ packages/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sylphx/silk-react
 
+## 1.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @sylphx/silk@1.2.0
+
 ## 1.0.2
 
 ### Patch Changes