+++ packages/vue-bindings/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @sylphx/silk-vue
 
+## 2.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @sylphx/silk@1.2.0
+
 ## 2.0.1
 
 ### Patch Changes